requirements.txt
PySide6>=6.5.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pillow-heif>=0.15.0
pyinstaller>=6.0.0 